+++ owlcms/src/main/java/app/owlcms/nui/results/ResultsContent.java
@@ -366,7 +366,7 @@ public class ResultsContent extends AthleteGridContent implements HasDynamicTitl
 
         // change the URL to reflect group
         event.getUI().getPage().getHistory().replaceState(null,
-                new Location(getLocation().getPath(), new QueryParameters(params)));
+                new Location(getLocation().getPath(), new QueryParameters(URLUtils.cleanParams(params))));
     }
 
     @Override
@@ -379,7 +379,7 @@ public class ResultsContent extends AthleteGridContent implements HasDynamicTitl
         } else {
             params.remove("group");
         }
-        ui.getPage().getHistory().replaceState(null, new Location(location.getPath(), new QueryParameters(params)));
+        ui.getPage().getHistory().replaceState(null, new Location(location.getPath(), new QueryParameters(URLUtils.cleanParams(params))));
     }
 
     @Override