+++ owlcms/src/main/java/app/owlcms/nui/shared/AthleteGridContent.java
@@ -1439,7 +1439,7 @@ public abstract class AthleteGridContent extends VerticalLayout
         } else {
             params.remove("group");
         }
-        ui.getPage().getHistory().replaceState(null, new Location(location.getPath(), new QueryParameters(params)));
+        ui.getPage().getHistory().replaceState(null, new Location(location.getPath(), new QueryParameters(URLUtils.cleanParams(params))));
     }
 
     /**