+++ owlcms/src/main/java/app/owlcms/nui/shared/BaseNavigationContent.java
@@ -163,7 +163,7 @@ public abstract class BaseNavigationContent extends VerticalLayout
         } else {
             params.remove("group");
         }
-        ui.getPage().getHistory().replaceState(null, new Location(location.getPath(), new QueryParameters(params)));
+        ui.getPage().getHistory().replaceState(null, new Location(location.getPath(), new QueryParameters(URLUtils.cleanParams(params))));
     }
 
     protected ComboBox<FieldOfPlay> createFopSelect(String placeHolder) {